modified:   MetaTradeNode/BlockchainService.h 	new file:   MetaTradeNode/LevelDBLocalImpl.cpp 	new file:   MetaTradeNode/LevelDBLocalImpl.h 	modified:   MetaTradeNode/LocalService.h 	modified:   MetaTradeNode/MetaTradeBlockchainImpl.cpp 	modified:   MetaTradeNode/MetaTradeBlockchainImpl.h 	modified:   MetaTradeNode/MetaTradeBlockchainModels.cpp 	modified:   MetaTradeNode/MetaTradeBlockchainModels.h 	deleted:    MetaTradeNode/MetaTradeLocalImpl.cpp 	deleted:    MetaTradeNode/MetaTradeLocalImpl.h 	new file:   MetaTradeNode/MetaTradeNode.cpp 	new file:   MetaTradeNode/MetaTradeNode.h 	new file:   MetaTradeNode/include/leveldb/c.h 	new file:   MetaTradeNode/include/leveldb/cache.h 	new file:   MetaTradeNode/include/leveldb/comparator.h 	new file:   MetaTradeNode/include/leveldb/db.h 	new file:   MetaTradeNode/include/leveldb/dumpfile.h 	new file:   MetaTradeNode/include/leveldb/env.h 	new file:   MetaTradeNode/include/leveldb/export.h 	new file:   MetaTradeNode/include/leveldb/filter_policy.h 	new file:   MetaTradeNode/include/leveldb/iterator.h 	new file:   MetaTradeNode/include/leveldb/options.h 	new file:   MetaTradeNode/include/leveldb/slice.h 	new file:   MetaTradeNode/include/leveldb/status.h 	new file:   MetaTradeNode/include/leveldb/table.h 	new file:   MetaTradeNode/include/leveldb/table_builder.h 	new file:   MetaTradeNode/include/leveldb/write_batch.h 	new file:   MetaTradeNode/include/port/port_config.h 	new file:   MetaTradeNode/lib/leveldb.lib 	new file:   dlls/leveldb.dll

+++ MetaTradeNode/MetaTradeBlockchainModels.h
@@ -5,8 +5,8 @@ namespace metatradenode {
 	struct Trade {
 		std::string senderAddress;
 		std::string receiverAddress;
-		double amount;
-		double commission;
+		long amount;
+		long commission;
 		long timestamp;
 		std::string signature;
 		std::string senderPublicKey;
